+++ notify-impl/src/main/java/com/dwarfeng/notify/impl/service/telqos/ResetCommand.java
@@ -2,18 +2,18 @@ package com.dwarfeng.notify.impl.service.telqos;
 
 import com.dwarfeng.notify.stack.handler.Resetter;
 import com.dwarfeng.notify.stack.service.ResetQosService;
+import com.dwarfeng.springtelqos.node.config.TelqosCommand;
 import com.dwarfeng.springtelqos.sdk.command.CliCommand;
 import com.dwarfeng.springtelqos.stack.command.Context;
 import com.dwarfeng.springtelqos.stack.exception.TelqosException;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.Option;
 import org.apache.commons.lang3.tuple.Pair;
-import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.List;
 
-@Component
+@TelqosCommand
 public class ResetCommand extends CliCommand {
 
     private static final String COMMAND_OPTION_LOOKUP = "l";